Add future.
Copy pipe mode to other file.
Add reset mode.
Edit mitlicense.hpp
