+++ tools/WinConsole/default.hpp
@@ -1,0 +1,25 @@
+#pragma once
+#include <tchar.h>
+const TCHAR *szDefatult_hostsfile=_T("\
+# Copyright (c) 1993-2009 Microsoft Corp.\n\
+#\n\
+# This is a sample HOSTS file used by Microsoft TCP/IP for Windows.\n\
+#\n\
+# This file contains the mappings of IP addresses to host names. Each\n\
+# entry should be kept on an individual line. The IP address should\n\
+# be placed in the first column followed by the corresponding host name.\n\
+# The IP address and the host name should be separated by at least one\n\
+# space.\n\
+#\n\
+# Additionally, comments (such as these) may be inserted on individual\n\
+# lines or following the machine name denoted by a '#' symbol.\n\
+#\n\
+# For example:\n\
+#\n\
+#      102.54.94.97     rhino.acme.com          # source server\n\
+#       38.25.63.10     x.acme.com              # x client host\n\
+\n\
+# localhost name resolution is handled within DNS itself.\n\
+#	127.0.0.1       localhost\n\
+#	::1             localhost\n\
+");
