+++ tools/WinConsole/tool.cpp
@@ -39,6 +39,8 @@
 #include "diff.hpp"
 #include "mitlicense.hpp"
 #include <signal.h>
+#include "pipedebug.hpp"
+#include "default.hpp"
 
 #define WIN32_LEAN_AND_MEAN
 
@@ -51,7 +53,7 @@
 #define objectwebsite _T("https://github.com/racaljk/hosts")
 //end.
 
-#define ConsoleTitle _T("racaljk-host tools     Build time:Apr. 26th, '16")
+#define ConsoleTitle _T("racaljk-host tools     Build time:Apr. 27th, '16")
 
 #define CASE(x,y) case x : y; break;
 #define pWait _T("\n    \
@@ -68,9 +70,10 @@ There seems something wrong in download file, we will retry after 5 seconds.\n")
 
 //debug set
 #define LogFileLocate _T("c:\\Hosts_Tool_log.log")
-#define pipeName _T("\\\\.\\pipe\\hoststoolnamepipe")
+const TCHAR * pipeName=_T("\\\\.\\pipe\\hoststoolnamepipe");
 #define szErrMeg _T("\nFatal Error:\n%s (GetLastError():%ld)\n\
 Please contact the application's support team for more information.\n")
+using namespace __Dpipe;
 //end.
 
 //for backward compatibility DO NOT CHANGE IT
@@ -90,28 +93,15 @@ struct expection{
 	}
 };
 
-//for pipe debug
-#define PIPE_TIMEOUT 5000
-#define BUFSIZE 4096
-typedef struct{
-	OVERLAPPED oOverlap;
-	HANDLE hPipeInst;
-	TCHAR chRequest[BUFSIZE];
-	DWORD cbRead;
-	TCHAR chReply[BUFSIZE];
-	DWORD cbToWrite;
-} PIPEINST, *LPPIPEINST;
-//end
-
 #define SHOW_HELP _T("\
 ------------------------------------------------------------\n\
 Hosts Tool for Windows Console by: Too-Naive\n\
 Copyright (C) 2016 @Too-Naive License:MIT LICENSE(redefined)\n\
 ------------------------------------------------------------\n\n\
-Usage: hosts_tool [-fi | -fu | -h | -? | -show | --debug-pipe]\n\n\
+Usage: hosts_tool [-? | -r | -fi | -fu | -show | --debug-pipe]\n\n\
 Options:\n\
-    -h    : Show this help message.\n\
     -?    : Show this help message.\n\
+    -r    : Reset system hosts file to default.\n\
     -fi   : Install Auto-Update hosts service(Service Name:%s).\n\
     -fu   : Uninstall service.\n\
     -show : Show the MIT license(redefined)\n\
@@ -138,7 +128,6 @@ SERVICE_STATUS ss;
 HANDLE lphdThread[]={
 	INVALID_HANDLE_VALUE,INVALID_HANDLE_VALUE
 };
-HANDLE hdPipe=INVALID_HANDLE_VALUE;
 bool request_client,bReserved;
 //end.
 
@@ -153,19 +142,8 @@ DWORD __stdcall NormalEntry(LPVOID);
 //DWORD __stdcall HostThread(LPVOID);
 void ___debug_point_reset(int);
 inline void __show_str(TCHAR const *,TCHAR const *);
-HANDLE ___pipeopen();
-inline DWORD ___pipeclose();
-DWORD __stdcall OpenPipeService(LPVOID);
-DWORD ___pipesentmessage(const TCHAR *);
+void Func_ResetFile();
 
-//pipe debug area
-void DisconnectAndClose(LPPIPEINST);
-BOOL CreateAndConnectInstance(LPOVERLAPPED);
-BOOL ConnectToNewClient(HANDLE, LPOVERLAPPED);
-inline void GetAnswerToRequest(LPPIPEINST);
-void WINAPI CompletedWriteRoutine(DWORD, DWORD, LPOVERLAPPED);
-void WINAPI CompletedReadRoutine(DWORD, DWORD, LPOVERLAPPED);
-//end.
 
 SERVICE_TABLE_ENTRY STE[2]={{Sname,Service_Main},{NULL,NULL}};
 
@@ -187,7 +165,7 @@ enum _Parameters{
 	DEBUG_SERVICE_START		=1<<0x09,
 	DEBUG_SERVICE_REINSTALL		=1<<0x0a,
 	OPEN_LISTEN			=1<<0x0b,
-	PARAMETERS_RESERVED5		=1<<0x0c,
+	RESET_FILE		=1<<0x0c,
 	PARAMETERS_RESERVED6		=1<<0x0d,
 	PARAMETERS_RESERVED7		=1<<0x0e,
 	PARAMETERS_RESERVED8		=1<<0x0f,
@@ -208,7 +186,8 @@ TCHAR const *szParameters[]={
 	_T("-debug-start"),			//8
 	_T("-debug-reiu"),			//9
 	_T("-debug-pipe"),			//10
-	_T("--pipe")				//11
+	_T("--pipe"),				//11
+	_T("r")					//12
 };
 
 int __fastcall __Check_Parameters(int argc,TCHAR const **argv){
@@ -234,6 +213,7 @@ int __fastcall __Check_Parameters(int argc,TCHAR const **argv){
 		case  8: return DEBUG_SERVICE_START;//start service
 		case  9: return DEBUG_SERVICE_REINSTALL;//reinstall service
 		case 10: return OPEN_LISTEN;
+		case 12: return RESET_FILE;
 		default: BAD_EXIT;
 	}
 	BAD_EXIT
@@ -249,14 +229,33 @@ int _tmain(int argc,TCHAR const ** argv){
 		CASE(EXEC_START_SERVICE,StartServiceCtrlDispatcher(STE));
 		CASE(EXEC_START_HELP,__show_str(SHOW_HELP,Sname));
 		CASE(EXEC_DEBUG_RESET,___debug_point_reset(EXEC_DEBUG_RESET));
-		CASE(SHOW_LICENSE,__show_str(szMitLicenseRaw,NULL));
+		CASE(SHOW_LICENSE,__show_str(szMitLicense_Raw,NULL));
 		CASE(DEBUG_SERVICE_STOP,___debug_point_reset(DEBUG_SERVICE_STOP));
 		CASE(DEBUG_SERVICE_START,___debug_point_reset(DEBUG_SERVICE_START));
 		CASE(DEBUG_SERVICE_REINSTALL,___debug_point_reset(DEBUG_SERVICE_REINSTALL));
 		CASE(OPEN_LISTEN,___debug_point_reset(OPEN_LISTEN));
+		CASE(RESET_FILE,Func_ResetFile());
 		default:break;
 	}
 	return 0;
+}
+
+void Func_ResetFile(){
+	_tprintf(_T("\
+------------------------------------------------------------\n\
+Hosts Tool for Windows Console by: Too-Naive\n\
+Copyright (C) 2016 @Too-Naive License:MIT LICENSE(redefined)\n\
+------------------------------------------------------------\n"));
+	if (!GetEnvironmentVariable(_T("SystemRoot"),buf3,BUFSIZ))
+		_tprintf(_T("    GetEnvironmentVariable() Error!(GetLastError():%ld)\n\
+\tCannot get system path!"),GetLastError()),abort();
+	_stprintf(buf1,_T("%s\\system32\\drivers\\etc\\hosts"),buf3);
+	FILE *fp=_tfopen(buf1,_T("w"));
+	if (!fp) _tprintf(_T("Cannot open file!\n")),abort();
+	_ftprintf(fp,_T("%s"),szDefatult_hostsfile);
+	fclose(fp);
+	_tprintf(_T("    Reset file successfully.\n"));
+	return ;
 }
 
 void __abrt(int){
@@ -457,34 +456,6 @@ inline void __fastcall ___checkEx(const TCHAR * szPstr,bool space_need){
 		___pipesentmessage(szPstr);
 }
 
-HANDLE ___pipeopen(){
-	while (1){
-		if ((hdPipe = CreateFile(pipeName,GENERIC_READ|GENERIC_WRITE,0,
-			NULL,OPEN_EXISTING,0,NULL))!=INVALID_HANDLE_VALUE)
-			break;
-		if (GetLastError()!=ERROR_PIPE_BUSY) {
-			Func_FastPMNTS(_T("%s Error! (%ld)\n"),__func__,GetLastError());
-			return INVALID_HANDLE_VALUE;
-		}
-		WaitNamedPipe(pipeName, 2000);
-	}
-	return hdPipe;
-}
-
-DWORD ___pipesentmessage(const TCHAR * szSent){
-	DWORD dwReserved=PIPE_READMODE_MESSAGE;
-    if (!SetNamedPipeHandleState(hdPipe,&dwReserved,NULL,NULL))
-		Func_FastPMNTS(_T("SetNamedPipeHandleState() Error! (%ld)\n"),GetLastError());
-    if (!WriteFile(hdPipe,szSent,(lstrlen(szSent)+1)*sizeof(TCHAR),&dwReserved,NULL))
-		Func_FastPMNTS(_T("WriteFile() Error! (%ld)\n"),GetLastError());
-    return GetLastError();
-}
-
-inline DWORD ___pipeclose(){
-    CloseHandle(hdPipe);
-	return GetLastError();
-}
-
 inline void __fastcall _perrtext(const TCHAR * _str,bool _Reserved){
 	if (!bReserved)	_tprintf(_str);
 	else if (_Reserved) Func_FastPMNTS(_str);
@@ -653,105 +624,3 @@ void WINAPI Service_Control(DWORD dwControl){
 	return ;
 }
 
-DWORD __stdcall OpenPipeService(LPVOID){
-	HANDLE hConnectEvent;
-	OVERLAPPED oConnect;
-	LPPIPEINST lpPipeInst;
-	DWORD dwWait, cbRet;
-	BOOL fSuccess, fPendingIO;
-	if (!(hConnectEvent = CreateEvent(NULL,TRUE,TRUE,NULL)))
-		return 0*_tprintf(_T("CreateEvent failed with %ld.\n"), GetLastError());
-	oConnect.hEvent = hConnectEvent;
-	fPendingIO = CreateAndConnectInstance(&oConnect);
-	while (1){
-		dwWait = WaitForSingleObjectEx(hConnectEvent,INFINITE,TRUE);
-		switch (dwWait){
-		case 0:
-			if (fPendingIO)
-				if (!(fSuccess = GetOverlappedResult(hdPipe,&oConnect,&cbRet,FALSE)))
-					return printf("ConnectNamedPipe (%ld)\n", GetLastError());
-			if (!(lpPipeInst=(LPPIPEINST) HeapAlloc(GetProcessHeap(),0,sizeof(PIPEINST))))
-				return printf("GlobalAlloc failed (%ld)\n", GetLastError());
-			lpPipeInst->hPipeInst = hdPipe;
-			lpPipeInst->cbToWrite = 0;
-			CompletedWriteRoutine(0, 0, (LPOVERLAPPED) lpPipeInst);
-			fPendingIO = CreateAndConnectInstance(&oConnect);
-			break;
-		case WAIT_IO_COMPLETION:
-			break;
-		default:
-			return printf("WaitForSingleObjectEx (%ld)\n", GetLastError());
-		}
-	}
-	return 0;
-}
-
-void WINAPI CompletedWriteRoutine(DWORD dwErr,DWORD cbWritten,LPOVERLAPPED lpOverLap){
-	LPPIPEINST lpPipeInst;
-	BOOL fRead = FALSE;
-	lpPipeInst = (LPPIPEINST) lpOverLap;
-	if ((dwErr == 0) && (cbWritten == lpPipeInst->cbToWrite)){
-		fRead = ReadFileEx(lpPipeInst->hPipeInst,lpPipeInst->chRequest,
-		BUFSIZE*sizeof(TCHAR),(LPOVERLAPPED) lpPipeInst,
-		(LPOVERLAPPED_COMPLETION_ROUTINE) CompletedReadRoutine);
-	}
-	if (!fRead) DisconnectAndClose(lpPipeInst);
-}
-
-void WINAPI CompletedReadRoutine(DWORD dwErr,DWORD cbBytesRead,LPOVERLAPPED lpOverLap){
-	LPPIPEINST lpPipeInst;
-	BOOL fWrite = FALSE;
-	lpPipeInst = (LPPIPEINST) lpOverLap;
-	if ((dwErr == 0) && (cbBytesRead != 0)){
-		GetAnswerToRequest(lpPipeInst);
-		fWrite = WriteFileEx(lpPipeInst->hPipeInst,lpPipeInst->chReply,
-			lpPipeInst->cbToWrite,(LPOVERLAPPED) lpPipeInst,
-			(LPOVERLAPPED_COMPLETION_ROUTINE) CompletedWriteRoutine);
-	}
-	if (!fWrite) DisconnectAndClose(lpPipeInst);
-}
-
-void DisconnectAndClose(LPPIPEINST lpPipeInst){
-	if (! DisconnectNamedPipe(lpPipeInst->hPipeInst))
-		printf("DisconnectNamedPipe failed with %ld.\n", GetLastError());
-	CloseHandle(lpPipeInst->hPipeInst);
-	if (lpPipeInst != NULL)
-		HeapFree(GetProcessHeap(),0, lpPipeInst);
-}
-
-BOOL CreateAndConnectInstance(LPOVERLAPPED lpoOverlap)
-{
-	if (!(hdPipe = CreateNamedPipe(pipeName,PIPE_ACCESS_DUPLEX |FILE_FLAG_OVERLAPPED,
-		PIPE_TYPE_MESSAGE |	PIPE_READMODE_MESSAGE |	PIPE_WAIT,
-		PIPE_UNLIMITED_INSTANCES,BUFSIZE*sizeof(TCHAR),BUFSIZE*sizeof(TCHAR),
-		PIPE_TIMEOUT,NULL)))
-		return 0*(printf("CreateNamedPipe failed with %ld.\n", GetLastError()));
-	return ConnectToNewClient(hdPipe, lpoOverlap);
-}
-
-BOOL ConnectToNewClient(HANDLE hPipe, LPOVERLAPPED lpo)
-{
-	BOOL fConnected, fPendingIO = FALSE;
-	if ((fConnected = ConnectNamedPipe(hPipe, lpo)))
-		return 0*printf("ConnectNamedPipe failed with %ld.\n", GetLastError());
-	switch (GetLastError()){
-	case ERROR_IO_PENDING:
-		fPendingIO = TRUE;
-		break;
-	case ERROR_PIPE_CONNECTED:
-		if (SetEvent(lpo->hEvent))
-			break;
-	default:
-			return 0*printf("ConnectNamedPipe failed with %ld.\n", GetLastError());
-	}
-	return fPendingIO;
-}
-
-inline void GetAnswerToRequest(LPPIPEINST pipe)
-{
-	_tprintf( TEXT("%s"), pipe->chRequest);
-	//reserved:
-/*	_tprintf( TEXT("[%p] %s"), pipe->hPipeInst, pipe->chRequest);
-	lstrcpyn( pipe->chReply,  TEXT("") ,BUFSIZE);
-	pipe->cbToWrite = (lstrlen(pipe->chReply)+1)*sizeof(TCHAR);*/
-}
